fixed alarm location (somewhat)
I adjusted the code to read the alarm .wav file from the C drive. I'd
like to have it read from the .jar file so that it's entirely portable,
but I'll settle for this Frankenstein-ed version for now. Next, I'll add
some additional code to restrict user input. I'm considering adding an
old-school rotating clock digits so that I have some practice with
graphics, but that's for later.