minor re-factoring, added more input restrictions
I added a second alarm input restriction based on whether or not the
alarm input's integers are within the reasonable range of a normal
24-hour clock. The input still does not test for whether or not the
alarm input is an integer, so that comes next. Afterward, I'll move on
to the more esoteric aspects. I plan to include some sort of randomized
action that must be completed. This will likely be some sort of string
that must be typed in or some sort of quick math equation with order of
operations that must be done. In the meantime, the alarm is blaring and
it doesn't turn off until the input is put in. Clever clever.